added style, changed capacity to percent indicator

+++ DataModels/FollowedEntityInfo.cs
@@ -11,7 +11,7 @@ namespace FirstPersonCameraContinued.DataModels
         public float currentSpeed;
         public int unitsSystem;
         public int passengers;
-        public int resources;
+        public float resources;
         public string? vehicleType;
         public string? citizenName;
         public string? citizenAction;
